REFACTOR(LectureItemDto): 이름만 다르고 내용이 중복되는 객체 통합
