+++ domain/src/main/java/com/capston/domain/response/lecture/DistinctLectureResponse.kt
@@ -1,7 +1,7 @@
 package com.capston.domain.response.lecture
 
 data class DistinctLectureResponse(
-    val data: List<LectureResponseDto>?,
+    val data: List<LectureItemDto>?,
     val nextCursor: Int = 0,
     val nextCreatedAt: String = "",
     val hasNext: Boolean = true
