+++ presentation/src/test/java/com/capston/presentation/viewmodel/PlanViewModelTest.kt
@@ -2,14 +2,13 @@ package com.capston.presentation.viewmodel
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.capston.domain.manager.LoadingStateManager
-import com.capston.domain.model.LectureItemDto
 import com.capston.domain.model.NewPlanLesson
 import com.capston.domain.request.LectureDto
 import com.capston.domain.response.enum_class.Platform
 import com.capston.domain.response.enum_class.Subject
 import com.capston.domain.response.lecture.DistinctLectureResponse
 import com.capston.domain.response.lecture.GetLessonsByLectureIdResponse
-import com.capston.domain.response.lecture.LectureResponseDto
+import com.capston.domain.response.lecture.LectureItemDto
 import com.capston.domain.usecase.lecture.GetAllLectureUseCase
 import com.capston.domain.usecase.lecture.GetDistinctLectureUseCase
 import com.capston.domain.usecase.lecture.GetLessonsByLectureIdUseCase
@@ -75,7 +74,7 @@ class PlanViewModelTest {
             subject = Subject.MATH
         )
         val lectures = listOf(
-            LectureResponseDto(
+            LectureItemDto(
                 id = 1,
                 title = "수학 기초",
                 teacher = "김선생",
@@ -105,7 +104,7 @@ class PlanViewModelTest {
         // Given
         val lectureDto = LectureDto(offset = "10")
         val lectures = listOf(
-            LectureResponseDto(
+            LectureItemDto(
                 id = 1,
                 title = "국어 문법",
                 teacher = "이선생",
@@ -113,7 +112,7 @@ class PlanViewModelTest {
                 subject = Subject.KOR,
                 totalLessons = 25
             ),
-            LectureResponseDto(
+            LectureItemDto(
                 id = 2,
                 title = "영어 독해",
                 teacher = "박선생",
@@ -253,7 +252,7 @@ class PlanViewModelTest {
             offset = "20"
         )
         val lectures = listOf(
-            LectureResponseDto(
+            LectureItemDto(
                 id = 1,
                 title = "영어 문법",
                 teacher = "김영어",
@@ -286,12 +285,12 @@ class PlanViewModelTest {
 
         val firstResponse = DistinctLectureResponse(
             data = listOf(
-                LectureResponseDto(id = 1, title = "수학 기초", teacher = "수학선생", platform = Platform.MEGA, subject = Subject.MATH)
+                LectureItemDto(id = 1, title = "수학 기초", teacher = "수학선생", platform = Platform.MEGA, subject = Subject.MATH)
             )
         )
         val secondResponse = DistinctLectureResponse(
             data = listOf(
-                LectureResponseDto(id = 2, title = "영어 기초", teacher = "영어선생", platform = Platform.ETOOS, subject = Subject.ENG)
+                LectureItemDto(id = 2, title = "영어 기초", teacher = "영어선생", platform = Platform.ETOOS, subject = Subject.ENG)
             )
         )
